+++ docker.py
@@ -22,6 +22,7 @@ DEFAULT_PROJECT_NAME = "mnist"
 DEFAULT_CODE_ROOT = "."
 DEFAULT_DATA_ROOT = "data"
 DEFAULT_LOG_ROOT = "log"
+DEFAULT_CACHE_ZSH_HISTORY = "./docker/misc/.zsh_history"
 
 
 class Env:
@@ -89,33 +90,26 @@ def execute(command):
         p.wait()
 
 
-def prepare_parser():
-    parser = argparse.ArgumentParser(
-        description="The core script of experiment management."
-    )
-    parser.add_argument("action", nargs="?", default="enter")
-    parser.add_argument("-b", "--build", action="store_true", default=False)
-    parser.add_argument("--root", action="store_true", default=False)
-
-    return parser
-
-
 def main(args):
-    _set_env(verbose=(args.action == "prepare"))
-    _prepare_mount_files()
+    _set_env(verbose=(args.action == "start" or args.action == "startd"))
 
-    service_name = "project"
-    if args.action == "prepare":
-        command = "docker-compose up -d"
+    SHELL = "zsh" if args.service == "project" else "bash"
+
+    if args.action == "start" or args.action == "startd":
+        command = "docker-compose up"
+        if args.action == "startd":
+            command += " -d"
         if args.build:
             command += " --build --force-recreate"
+        command += f" {args.service}"
     elif args.action == "enter":
         if args.root:
-            command = f"docker-compose exec -u root {service_name} zsh"
+            command = f"docker-compose exec -u root {args.service} {SHELL}"
         else:
-            command = f"docker-compose exec {service_name} zsh"
+            command = f"docker-compose exec {args.service} {SHELL}"
     else:
         command = f"docker-compose {args.action}"
+    print(f"> {command}\n")
     execute(command)
 
 
@@ -173,22 +167,41 @@ def _set_env(env_path=DEFAULT_ENV_PATH, verbose=False):
                 log_root.mkdir(parents=True)
         e["LOG_ROOT"] = str(log_root)
 
-    e["COMPOSE_PROJECT_NAME"] = f"{e['PROJECT']}_{e['USER_NAME']}"
+    if "CACHE_ZSH_HISTORY" not in e:
+        cache_zsh_history = Path(
+            _get_value_from_stdin(
+                "file to be synced with ~/.zsh_history",
+                default=DEFAULT_CACHE_ZSH_HISTORY,
+            )
+        ).resolve()
+        if not cache_zsh_history.exists():
+            if (
+                _get_value_from_stdin(
+                    f"`{cache_zsh_history}` does not exist in your machine. Create?",
+                    default="yes",
+                )
+                == "yes"
+            ):
+                cache_zsh_history.touch()
+        e["CACHE_ZSH_HISTORY"] = str(cache_zsh_history)
+
+    e["COMPOSE_PROJECT_NAME"] = f"{e['PROJECT']}_{e['USER_NAME']}".lower()
     e.save()
 
     if verbose:
-        print(f"Your setting ({env_path}):\n{e}")
+        print(f"Your setting ({env_path}):\n{e}\n")
     return e
 
 
-def _prepare_mount_files():
-    PATH_ZSH_HISTORY = Path("./docker/misc/.zsh_history")
-    if not PATH_ZSH_HISTORY.exists():
-        PATH_ZSH_HISTORY.touch()
-        print(f"{PATH_ZSH_HISTORY} is created.")
-
-
 if __name__ == "__main__":
-    parser = prepare_parser()
+    parser = argparse.ArgumentParser(
+        description="The core script of experiment management."
+    )
+    parser.add_argument("action", nargs="?", default="enter")
+    parser.add_argument("-b", "--build", action="store_true", default=False)
+    parser.add_argument("--root", action="store_true", default=False)
+    parser.add_argument("--service", default="project")
+
     args = parser.parse_args()
+
     main(args)