+++ scripts/testing/test_lm.sh
@@ -1,0 +1,17 @@
+# ViT-B/32, 2 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-12 --update_wandb
+
+# ViT-B/16, 2 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-13 --update_wandb
+
+# ViT-L/14, 2 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-14 --update_wandb
+
+# ViT-B/32, 4 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-15 --update_wandb
+
+# ViT-B/16, 4 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-16 --update_wandb
+
+# ViT-L/14, 4 layers
+python test.py /log/exp/vtt/WikiHowDataModule.CLIPAutoEncoder.GenerationLoss.2023-01-14_00-11-17 --update_wandb