+++ src/dataset/text.py
@@ -4,9 +4,11 @@ import gzip
 import html
 import os
 from functools import lru_cache
+from typing import List, Union
 
 import ftfy
 import regex as re
+import torch
 
 
 @lru_cache()
@@ -72,7 +74,8 @@ def whitespace_clean(text):
 
 
 class SimpleTokenizer(object):
-    def __init__(self, bpe_path: str = default_bpe()):
+    def __init__(self, bpe_path: str = default_bpe(), max_words: int = 77):
+        self.max_words = max_words
         self.byte_encoder = bytes_to_unicode()
         self.byte_decoder = {v: k for k, v in self.byte_encoder.items()}
         merges = gzip.open(bpe_path).read().decode("utf-8").split("\n")
@@ -182,3 +185,60 @@ class SimpleTokenizer(object):
                 return self.decode(tokens[:i])
         else:
             return ""
+
+    def tokenize(
+        self,
+        texts: Union[str, List[str]],
+        truncate: bool = False,
+        return_mask: bool = False,
+    ) -> torch.LongTensor:
+        """Returns the tokenized representation of given input string(s)
+
+        Parameters
+        ----------
+        texts : Union[str, List[str]]
+            An input string or a list of input strings to tokenize
+
+        context_length : int
+            The context length to use; all CLIP models use 77 as the context length
+
+        truncate: bool
+            Whether to truncate the text in case its encoding is longer than the context length
+
+        Returns
+        -------
+        A two-dimensional tensor containing the resulting tokens, shape = [number of input strings, context_length]
+        """
+        if isinstance(texts, str):
+            texts = [texts]
+
+        context_length = self.max_words
+
+        sot_token = self.encoder["<|startoftext|>"]
+        eot_token = self.encoder["<|endoftext|>"]
+        all_tokens = [
+            [sot_token] + self.encode(text) + [eot_token] for text in texts
+        ]
+        result = torch.zeros(len(all_tokens), context_length, dtype=torch.long)
+        if return_mask:
+            mask = torch.zeros(
+                len(all_tokens), context_length, dtype=torch.bool
+            )
+
+        for i, tokens in enumerate(all_tokens):
+            if len(tokens) > context_length:
+                if truncate:
+                    tokens = tokens[:context_length]
+                    tokens[-1] = eot_token
+                else:
+                    raise RuntimeError(
+                        f"Input {texts[i]} is too long for context length {context_length}"
+                    )
+            result[i, : len(tokens)] = torch.tensor(tokens)
+            if return_mask:
+                mask[i, : len(tokens)] = True
+
+        if return_mask:
+            return result, mask
+        else:
+            return result